Share failure cleanup chains end to end

+++ crates/lkjscript-ir/src/model/ids.rs
@@ -28,7 +28,7 @@ dense_id!(TraitId, u32);
 dense_id!(ImplId, u32);
 dense_id!(PlaceId, u32);
 dense_id!(LoanId, u32);
-dense_id!(FailureCleanupId, u32);
+dense_id!(FailureCleanupId, u64);
 dense_id!(StructuralTypeId, u16);
 dense_id!(StructuralLayoutId, u16);
 dense_id!(StructuralRepresentationId, u16);
